Move `target.py` rules to `engine/internals/graph.py`
[ci skip-rust-tests]
